Update authentication tests.
This updates some of the remaining authentication tests to use a mocked
version of the fed auth flow. I've also merged them into the
tests/routes/test_auth.py file, which is a more appropriate home for
them.
